+++ src/hybran/BLAST.py
@@ -22,11 +22,41 @@ blast_prog = {
     'p': 'blastp',
 }
 
-def reference_match(query, subject, seq_ident, seq_covg, identify=lambda _:_, metric='iden', blast_type = "p", strict=False):
+presets = {
+    'thorough': [
+        # comp_based_stats described in:
+        # Yu & Altschul, 2005 doi:10.1093/bioinformatics/bti070
+        # Reason for this setting:
+        # https://github.com/ncbi/blast_plus_docs/issues/23
+        '-comp_based_stats', '0',
+        '-seg', 'no',
+    ],
+    'conservative': [
+        '-comp_based_stats', '2',
+        # Moreno-Hagelsieb & Latimer, 2008 doi:10.1093/bioinformatics/btm585
+        #'-use_sw_tback', # authors said this doesn't add much improvement but slows things down signicantly
+        '-seg', 'yes',
+        '-soft_masking', 'true',
+    ],
+}
+
+
+def reference_match(
+        query,
+        subject,
+        cutoff=config.cnf.blast.min_identity,
+        seq_covg=config.cnf.blast.min_coverage,
+        identify=lambda _:_,
+        metric='iden',
+        blast_type="p",
+        strict=False,
+        preset='thorough',
+):
     """
     Wrapper to blast(), summarize(), and top_hit() that tells you the bottom line.
     See those functions for descriptions of the main input parameters.
     This function checks for complete hits and, failing that, partial hits.
+    :param query: SeqRecord a single gene to search
     :param strict: Boolean whether to strictly enforce all thresholds or allow for low-coverage alignments.
     :returns:
         - result (:py:class:`str`)  -
@@ -38,13 +68,15 @@ def reference_match(query, subject, seq_ident, seq_covg, identify=lambda _:_, me
         - stats (:py:class:`dict`)  -
              summary dict of blast results from either hits, truncation_signatures, or misses
     """
-    
+
     hits, misses, truncation_signatures = blast(
-        query,
-        subject,
-        seq_ident,
-        seq_covg,
-        blast_type,
+        query=query,
+        subject=subject,
+        metric=metric,
+        cutoff=cutoff,
+        seq_covg=seq_covg,
+        blast_type=blast_type,
+        blast_extra_args=presets[preset],
     )
     result = None
     low_covg = False
@@ -56,12 +88,12 @@ def reference_match(query, subject, seq_ident, seq_covg, identify=lambda _:_, me
         principal_lists = [hits, truncation_signatures]
         scraps = misses
     # this will be replaced by actual hits later if there are any
-    hit_dict = summarize(scraps, identify=identify)
+    hit_dict = summarize(scraps, identify=identify)[query.id]
 
     for hit_list in principal_lists:
         if not hit_list:
             continue
-        hit_dict = summarize(hit_list, identify=identify)
+        hit_dict = summarize(hit_list, identify=identify)[query.id]
         result = top_hit(
             hit_dict,
             metric=metric,
@@ -69,9 +101,11 @@ def reference_match(query, subject, seq_ident, seq_covg, identify=lambda _:_, me
         stats = hit_dict[result]
         scov_pass = stats['scov'] >= seq_covg
         qcov_pass = stats['qcov'] >= seq_covg
-        ident_pass = stats['iden'] >= seq_ident
-        if ident_pass and any([scov_pass, qcov_pass]) and not all([scov_pass, qcov_pass]):
-            low_covg = True
+        low_covg = (
+            stats[metric] >= cutoff
+            and any([scov_pass, qcov_pass])
+            and not all([scov_pass, qcov_pass])
+        )
         # if we found a proper hit, don't look for anything else
         break
 
@@ -87,6 +121,19 @@ def top_hit(blast_summary, metric='iden'):
     """
     return max(blast_summary, key=lambda gene: blast_summary[gene][metric])
 
+def top_hits(blast_summary, metric='iden'):
+    """
+    Get the best hits according to the given metric, those with scores
+    >= 99% of the highest absolute score.
+    See Wolf & Koonin, 2012 (doi:10.1093/gbe/evs100)
+    """
+    top_score = max([blast_summary[gene][metric] for gene in blast_summary])
+    return sorted(
+        [gene for gene in blast_summary if blast_summary[gene][metric] >= 0.99 * top_score],
+        key=lambda gene: blast_summary[gene][metric],
+        reverse=True,
+    )
+
 def summarize(blast_results, identify = lambda _:_):
     """
     Convert output of blastp() into a dictionary of the following form:
@@ -101,36 +148,54 @@ def summarize(blast_results, identify = lambda _:_):
     :returns: dictionary of subject IDs to metrics to values that
               defaults to a null hit for missing keys.
     """
-    summary = defaultdict(lambda : {
+    summary = defaultdict(lambda: defaultdict(lambda : {
         'iden':0.,
+        #'evalue': 1e10,
+        'bitscore':0.,
         'qcov':0.,
         'scov':0.,
-    })
+    }))
     for line in reversed(blast_results):
         fields = line.split('\t')
-        summary[identify(fields[1])] = {
+        summary[identify(fields[0])][identify(fields[1])] = {
             'iden':float(fields[2]),
+            #'evalue': float(fields[10]),
+            'bitscore': float(fields[11]),
             'qcov':float(fields[14]),
             'scov':float(fields[15]),
         }
+
     return summary
 
-def blastn(query, subject, seq_ident, seq_covg):
-        return blast(query, subject, seq_ident, seq_covg, blast_type = "n")
+def blastn(query, subject, metric, cutoff, seq_covg, nproc=1, blast_extra_args=None, screen=True):
+        return blast(query, subject, metric, cutoff, seq_covg, blast_type="n", nproc=nproc, blast_extra_args=blast_extra_args, screen=screen)
         
-def blastp(query, subject, seq_ident, seq_covg):
-        return blast(query, subject, seq_ident, seq_covg, blast_type = "p")
+def blastp(query, subject, metric, cutoff, seq_covg, nproc=1, blast_extra_args=None, screen=True):
+        return blast(query, subject, metric, cutoff, seq_covg, blast_type="p", nproc=nproc, blast_extra_args=blast_extra_args, screen=screen)
 
-def blast(query, subject, seq_ident, seq_covg, blast_type = "p"):
+def blast(
+        query,
+        subject,
+        metric,
+        cutoff,
+        seq_covg,
+        blast_type="p",
+        nproc=1,
+        blast_extra_args=None,
+        screen=True,
+):
     """
-    Runs BLAST with one sequence as the query and
-    the fasta file as the subject.
+    Runs BLAST
 
-    :param query: SeqRecord object containing the query sequence/id/description
+    :param query: str query fasta file name or SeqRecord object containing the query sequence/id/description
     :param subject: str subject fasta file name or SeqRecord
-    :param seq_ident: sequence identity threshold
+    :param metric: str name of parameter to use as primary threshold (in addition to alignment coverage)
+    :param cutoff: float threshold for selected metric
     :param seq_covg: alignment coverage threshold
     :param blast_type: nucleotide or protein blast
+    :param nproc: int number of threads to use for blast
+    :param blast_extra_args: list of str tokens to pass to blast executable
+    :param screen: bool whether to apply thresholds/filtering to results
     :returns:
         - blast_filtered (:py:class:`list`) - list of tab-delimited strings corresponding
                                               to BLAST hits meeting the thresholds
@@ -141,6 +206,13 @@ def blast(query, subject, seq_ident, seq_covg, blast_type = "p"):
 					      meet the threshold, while the remaining coverage factor fails.
 
     """
+    # prevent a mutable variable (list) being defined in the function header
+    if blast_extra_args is None:
+        blast_extra_args = []
+    # prevent changes to the argument (which is usually the preset list) from persisting
+    else:
+        blast_extra_args = blast_extra_args.copy()
+
     if isinstance(subject, SeqRecord):
         subject_id = subject.id
         with tempfile.NamedTemporaryFile(
@@ -163,6 +235,20 @@ def blast(query, subject, seq_ident, seq_covg, blast_type = "p"):
             subject_id = rec.id
             count += 1
 
+    if not isinstance(query, SeqRecord):
+        blast_extra_args += ['-query', query]
+        blast_stdin = None
+        count = 1
+        for rec in SeqIO.parse(query,'fasta'):
+            if count > 1:
+                query_id = ''
+                break
+            query_id = rec.id
+            count += 1
+    else:
+        query_id = query.id
+        blast_stdin = str(query.seq)
+
     # column orders up until bitscore are as expected for mcxdeblast --m9
     # <https://github.com/JohannesBuchner/mcl/blob/5208b974324621f510abb6a29e046a38f6d85f10/src/alien/oxygen/src/mcxdeblast#L259>
     blast_outfmt = "6 qseqid sseqid pident length mismatch gapopen " \
@@ -170,7 +256,7 @@ def blast(query, subject, seq_ident, seq_covg, blast_type = "p"):
     # for a single alignment, add a dummy hit with 0% identity so that 0-thresholds are valid
     # when there are no hits at all.
     if subject_id:
-        dummy_hit = '\t'.join([query.id, subject_id,'0','0','0','0',
+        dummy_hit = '\t'.join([query_id, subject_id,'0','0','0','0',
                                '','','','','1','0','1','1','',''])+'\n'
     else:
         dummy_hit = ''
@@ -178,74 +264,133 @@ def blast(query, subject, seq_ident, seq_covg, blast_type = "p"):
         blast_prog[blast_type],
         '-subject', fa,
         '-outfmt', blast_outfmt,
-    ]
+        '-num_threads', str(nproc),
+    ] + blast_extra_args
     blast_ps = subprocess.run(
         blast_cmd,
-        input=str(query.seq),
+        input=blast_stdin,
         text=True,
         check=True,
         capture_output=True,
     )
+
+    if not screen:
+        return blast_ps.stdout.split('\n'), [], []
+
     stdout = blast_ps.stdout + dummy_hit
     blast_filtered = []
     blast_rejects = []
     blast_truncation = []
     for line in stdout.split('\n'):
-        if line:
-            column = line.split('\t')
-            identity = float(column[2])
-            length = float(column[3])
-            qlen = float(column[12])
-            slen = float(column[13])
-            qseq, sseq = column[14:16]
-            qcov = (len(qseq.replace("-","")) / qlen) * 100
-            scov = (len(sseq.replace("-","")) / slen) * 100
-            column[14:16] = str(qcov), str(scov)
+        if not line:
+            continue
+        values = {}
+
+        column = line.split('\t')
+        values['iden'] = float(column[2])
+        length = float(column[3])
+        values['bitscore'] = float(column[11])
+        qlen = float(column[12])
+        slen = float(column[13])
+        qseq, sseq = column[14:16]
+        qcov = (len(qseq.replace("-","")) / qlen) * 100
+        scov = (len(sseq.replace("-","")) / slen) * 100
+        column[14:16] = str(qcov), str(scov)
+        if blast_stdin is not None:
             column[0] = query.id
-            if (identity >= seq_ident) and (qcov >= seq_covg) and (scov >= seq_covg):
-                blast_filtered.append('\t'.join(column))
-            elif (identity >= seq_ident) and (qcov >= seq_covg) and (scov <= seq_covg):
-                blast_truncation.append('\t'.join(column))
-            elif (identity >= seq_ident) and (qcov <= seq_covg) and (scov >= seq_covg):
-                blast_truncation.append('\t'.join(column))
-            else:
-                # Capture the rejected hits
-                blast_rejects.append('\t'.join(column))
+        if (values[metric] >= cutoff) and (qcov >= seq_covg) and (scov >= seq_covg):
+            blast_filtered.append('\t'.join(column))
+        elif (values[metric] >= cutoff) and (qcov >= seq_covg) and (scov <= seq_covg):
+            blast_truncation.append('\t'.join(column))
+        elif (values[metric] >= cutoff) and (qcov <= seq_covg) and (scov >= seq_covg):
+            blast_truncation.append('\t'.join(column))
+        else:
+            # Capture the rejected hits
+            blast_rejects.append('\t'.join(column))
 
     return blast_filtered, blast_rejects, blast_truncation
 
 
-def iterate(fa, seq_list, nproc, seq_ident, seq_covg):
-    """
-    Runs BLAST function for each query.
-    Returns list of all results.
-    """
-    partial_blast = partial(blastp, subject=fa, seq_ident=seq_ident, seq_covg=seq_covg)
-    pool = multiprocessing.Pool(int(nproc))
-    hits, misses, truncation_signatures = zip(*pool.map(partial_blast,seq_list))
-    pool.close()
-    pool.join()
-    all_results_list = list(map('\n'.join, list(hits)))
-    return all_results_list
-
-
-def write(all_results_list):
-    """
-    Joins list of all results into a continuous
-    string and writes this string to a new file
-    """
-    hybran_tmp_dir = config.hybran_tmp_dir
-    joined_string = '\n'.join(all_results_list)
-    with open(hybran_tmp_dir + '/blast_results', 'w') as all_v_all:
-        all_v_all.write(joined_string)
-    return all_v_all
-
-
-def run_blast(fastafile, nproc, seq_ident, seq_covg):
+def all_vs_all(
+        fastafile,
+        nproc,
+        metric='iden',
+        cutoff=config.cnf.blast.min_identity,
+        seq_covg=config.cnf.blast.min_coverage,
+        blast_preset='thorough',
+):
     logger = logging.getLogger('BLAST')
     logger.info('Running pairwise all-against-all BLAST on ' + fastafile + ' using ' + str(nproc) + ' CPUs')
-    seq_string_list = SeqIO.parse(fastafile, 'fasta')
-    all_results_list = iterate(fastafile, seq_string_list, nproc, seq_ident=seq_ident, seq_covg=seq_covg)
+
+    all_results, _, _ = blastp(
+        query=fastafile,
+        subject=fastafile,
+        metric=metric,
+        cutoff=cutoff,
+        seq_covg=seq_covg,
+        nproc=nproc,
+        blast_extra_args=presets[blast_preset],
+        screen=True,
+    )
+
     logger.info('Writing BLAST results to blast_results in Hybran temporary '
                 'directory.')
-    write(all_results_list)
+    hybran_tmp_dir = config.hybran_tmp_dir
+    with open(hybran_tmp_dir + '/blast_results', 'w') as all_v_all:
+        all_v_all.write('\n'.join(all_results))
+
+def bidirectional_best_hit(
+        query,
+        subject,
+        metric='iden',
+        cutoff=config.cnf.blast.min_identity,
+        min_seq_covg=config.cnf.blast.min_coverage,
+        identify=lambda _:_,
+        nproc=1,
+        blast_type="p",
+):
+    bbh_results = defaultdict(lambda :None)
+    blast_extra_args = presets['conservative']
+
+    qry2ref_hits, _, _ = blast(
+        query,
+        subject,
+        metric=metric,
+        cutoff=cutoff,
+        seq_covg=min_seq_covg,
+        blast_type=blast_type,
+        nproc=nproc,
+        blast_extra_args=blast_extra_args,
+    )
+    qry2ref_hit_dict = summarize(qry2ref_hits, identify=identify)
+    ref2qry_hits, _, _ = blast(
+        subject,
+        query,
+        metric=metric,
+        cutoff=cutoff,
+        seq_covg=min_seq_covg,
+        blast_type=blast_type,
+        nproc=nproc,
+        blast_extra_args=blast_extra_args,
+    )
+    ref2qry_hit_dict = summarize(ref2qry_hits, identify=identify)
+    ref2qry_top_hits = defaultdict(lambda : [])
+    ref2qry_top_hits.update({
+        ref_gene:top_hits(ref_hit_summary)
+        for ref_gene, ref_hit_summary in ref2qry_hit_dict.items()
+    })
+
+    for query_gene in qry2ref_hit_dict:
+        bbh_match = None
+        score_to_beat = 0.
+        qry_top_hits = top_hits(qry2ref_hit_dict[query_gene])
+        for ref_gene in qry_top_hits:
+            if (
+                    query_gene in ref2qry_top_hits[ref_gene]
+                    and qry2ref_hit_dict[query_gene][ref_gene][metric] > score_to_beat
+            ):
+                bbh_match = ref_gene
+                score_to_beat = qry2ref_hit_dict[query_gene][ref_gene][metric]
+        bbh_results[query_gene] = bbh_match
+
+    return bbh_results, qry2ref_hit_dict