+++ src/hybran/main.py
@@ -31,6 +31,12 @@ warnings.filterwarnings(
     message=r"Partial codon",
 )
 
+def percentage(string):
+    num = float(string)
+    if not (num <= 100 and num >= 0):
+        raise argparse.ArgumentTypeError("Not a percentage.")
+    return num
+
 def cmds():
     """
     argparse parse input provided by the user and call appropriate command
@@ -156,16 +162,16 @@ def cmds():
     onegenecmd.add_argument(
         '-i', '--identity-threshold',
         required=False,
-        type=int,
-        help='Percent sequence identity threshold to use during CD-HIT clustering and BLASTP to call duplications',
-        default=99,
+        type=percentage,
+        help='Percent sequence identity threshold to use for considering sequences as redundant.',
+        default=config.cnf.onegene.min_identity,
     )
     onegenecmd.add_argument(
         '-c', '--coverage-threshold',
         required=False,
-        type=int,
-        help='Percent alignment coverage threshold to use during CD-HIT clustering and BLASTP to call duplications',
-        default=99,
+        type=percentage,
+        help='Percent alignment coverage threshold to use for considering sequences as redundant.',
+        default=config.cnf.onegene.min_coverage,
     )
     onegenecmd.add_argument(
         '-t', '--first-reference',
@@ -210,8 +216,21 @@ def cmds():
     #
     required = parser.add_argument_group('Required')
     optional = parser.add_argument_group('Optional')
-    ratt_params = parser.add_argument_group('RATT Options.\n(See http://ratt.sourceforge.net/documentation.html and\n https://github.com/ThomasDOtto/ratt for more details)')
-    prokka_params = parser.add_argument_group('Prokka Options.\n(See https://github.com/tseemann/prokka for more details)')
+    main_params = parser.add_argument_group(
+        'Main Parameters Affecting the Inference of Gene Homologs',
+        description=(
+            'BLAST is used (for bidirectional best hits) in matching ab initio predicted genes to reference genes,'
+            ' as well as in preparing the input all vs. all hit matrix for MCL.'
+        ),
+    )
+    ratt_params = parser.add_argument_group(
+        'RATT Options',
+        description='See <https://ratt.sourceforge.net/documentation.html> and <https://github.com/ThomasDOtto/ratt> for more details.',
+    )
+    prokka_params = parser.add_argument_group(
+        'Prokka Options',
+        description='See https://github.com/tseemann/prokka for more details.',
+    )
     required.add_argument('-g', '--genomes', help='Directory, a space-separated list of FASTAs, or a FOFN '
                                                       'containing all genomes desired to be annotated. '
                                                       'FASTA format required.',
@@ -234,12 +253,6 @@ def cmds():
     optional.add_argument('-t', '--first-reference', required=False, dest='first_gbk',
                           help="Reference name or file name whose locus tags should be used as unified names for conserved copies in the others."
                           " Default is the annotation with the most named CDSs. If you specify a file here that is not in your input list, it will be added.")
-    optional.add_argument('-i', '--identity-threshold', required=False, type=int,
-                          help='Percent sequence identity threshold to use during CD-HIT clustering and BLASTP',
-                          default=95)
-    optional.add_argument('-c', '--coverage-threshold', required=False, type=int,
-                          help='Percent sequence coverage threshold to use during CD-HIT clustering and BLASTP',
-                          default=95)
     optional.add_argument('-o', '--output', help='Directory to output all new annotation files.',
                           default='.')
     optional.add_argument('-n', '--nproc', help='Number of processors/CPUs to use',
@@ -263,6 +276,44 @@ def cmds():
                               'This option is deprecated and has no effect.'
                               'If name unification is not desired, consider running `hybran standardize` afterwards.'
                           ))
+
+    main_params.add_argument(
+        '--onegene-identity-threshold',
+        help='Minimum percent sequence identity threshold to use for identifying redundant sequences.',
+        required=False,
+        type=percentage,
+        default=config.cnf.onegene.min_identity,
+    )
+    main_params.add_argument(
+        '--onegene-coverage-threshold',
+        help='Minimum percent sequence coverage threshold to use for identifying redundant sequences.',
+        required=False,
+        type=percentage,
+        default=config.cnf.onegene.min_coverage,
+    )
+    main_params.add_argument(
+        '-i', '--blast-min-identity', '--identity-threshold',
+        help="Minimum percent sequence identity threshold to use for inferring homologs.",
+        default=config.cnf.blast.min_identity,
+        type=percentage,
+    )
+    main_params.add_argument(
+        '-c', '--blast-min-coverage', '--coverage-threshold',
+        help="Minimum percent query and subject alignment coverage threshold to use for inferring homologs.",
+        default=config.cnf.blast.min_coverage,
+        type=percentage,
+    )
+    main_params.add_argument(
+        '-I', '--mcl-inflation',
+        help=(
+            "MCL inflation value. "
+            "Higher value results in more fine-grained clusters (fewer genes in common). "
+            "See <https://micans.org/mcl/man/mcl.html#opt-I> for details."
+        ),
+        type=float,
+        default=config.cnf.mcl_inflation,
+    )
+
 
     logging_level = optional.add_mutually_exclusive_group()
     logging_level.add_argument('--verbose', action='store_true', help='Verbose output')
@@ -377,6 +428,11 @@ def main(args, prokka_args):
     # Setting up the Hybran temporary directory
     config.init()
     hybran_tmp_dir = config.hybran_tmp_dir
+    config.cnf.onegene.min_identity =  args.onegene_identity_threshold
+    config.cnf.onegene.min_coverage =  args.onegene_coverage_threshold
+    config.cnf.mcl_inflation = args.mcl_inflation
+    config.cnf.blast.min_coverage = args.blast_min_coverage
+    config.cnf.blast.min_identity = args.blast_min_identity
 
     designator.generic_orf_prefix[0]=args.orf_prefix
     designator.ref_orf_prefix[0] = f"REF{args.orf_prefix}X"
@@ -385,16 +441,6 @@ def main(args, prokka_args):
     # --debug is set
     if not args.debug:
         atexit.register(shutil.rmtree, path=hybran_tmp_dir)
-
-    # Check that the identity threshold is valid
-    if not (args.identity_threshold <= 100 and args.identity_threshold >= 0):
-        print("error: invalid value for --identity-threshold. Must be between 0 and 100.")
-        exit(10)
-
-    # Check that the coverage threshold is valid
-    if not (args.coverage_threshold <= 100 and args.coverage_threshold >= 0):
-        print("error: invalid value for --coverage-threshold. Must be between 0 and 100.")
-        exit(10)
 
     # Confirming all installations are valid
     if args.database_dir:
@@ -539,7 +585,7 @@ This option is scheduled for removal, so please update your invocation for the f
                                 prokka_extra_args = prokka_args,
                                 script_dir=script_dir,
                                 cpus=args.nproc,
-                                qcov=args.coverage_threshold)
+                                qcov=config.cnf.blast.min_coverage)
                 if not os.path.isfile(annomerge_gbk):
                     logger.info('Merging RATT and Prokka annotations for ' + samplename)
                     annomerge.run(isolate_id=samplename,
@@ -550,8 +596,6 @@ This option is scheduled for removal, so please update your invocation for the f
                                   ref_proteins_fasta=ref_cds,
                                   ref_gbk_list=ref_gbks,
                                   script_directory=script_dir,
-                                  seq_ident=args.identity_threshold,
-                                  seq_covg=args.coverage_threshold,
                                   ratt_enforce_thresholds=args.filter_ratt,
                                   nproc=args.nproc,
                     )
@@ -564,11 +608,11 @@ This option is scheduled for removal, so please update your invocation for the f
                 shutil.copy(annomerge_gff, args.output)
     if all([os.path.isfile(os.path.join(args.output, os.path.basename(g) + '.gff')) for g in genomes]):
         all_genomes += [refdir + i for i in os.listdir(refdir) if i.endswith('.gff')] + genomes_annotate
-        run.clustering(all_genomes=list(set(sorted(all_genomes))),
-                       target_genomes=genomes_annotate,
-                       nproc=args.nproc,
-                       seq_ident=args.identity_threshold,
-                       seq_covg=args.coverage_threshold)
+        run.clustering(
+            all_genomes=list(set(sorted(all_genomes))),
+            target_genomes=genomes_annotate,
+            nproc=args.nproc,
+        )
         if args.database_dir and 'eggnog_seqs.fasta' in os.listdir(hybran_tmp_dir):
             ref_tax_ids = [extractor.get_taxonomy_id(_) for _ in ref_gbks]
             if any(ref_tax_ids):