+++ src/hybran/prokka.py
@@ -3,6 +3,7 @@ import functools
 import logging
 import os
 import re
+import tempfile
 
 from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
@@ -36,8 +37,6 @@ def postprocess(
         postprocess_outdir,
         ref_annotation,
         ref_proteome,
-        seq_ident,
-        seq_covg,
         nproc=1,
 ):
 
@@ -76,8 +75,6 @@ def postprocess(
             record=record,
             ref_annotation=ref_annotation,
             ref_proteome=ref_proteome,
-            seq_ident=seq_ident,
-            seq_covg=seq_covg,
             nproc=nproc,
         )
         prokka_features[contigs[i]] = record.features
@@ -109,8 +106,6 @@ def postprocess_contig(
         record,
         ref_annotation,
         ref_proteome,
-        seq_ident,
-        seq_covg,
         nproc=1,
 ):
 
@@ -134,14 +129,31 @@ def postprocess_contig(
     abinit_blast_results_complete = {}
     # only contains results for the accepted reference gene hit
     abinit_blast_results = {}
+    prokka_contig_cdss = [f for f in contig_features if f.type == 'CDS']
+    with tempfile.NamedTemporaryFile(
+            suffix='.fasta',
+            dir=config.hybran_tmp_dir,
+            delete=False,
+            mode='w',
+    ) as fa_handle:
+        cds_records = [
+            SeqRecord(Seq(f.qualifiers['translation'][0]), id=f.qualifiers['locus_tag'][0])
+            for f in prokka_contig_cdss
+        ]
+        cds_fasta = fa_handle.name
+        SeqIO.write(cds_records, cds_fasta, "fasta")
+    bbh_results, all_qry2ref_blast_hits = BLAST.bidirectional_best_hit(
+        cds_fasta,
+        ref_proteome,
+        nproc=nproc,
+    )
     mp_postprocess_feature = functools.partial(
         postprocess_feature,
         ref_annotation=ref_annotation,
         ref_proteome=ref_proteome,
-        seq_ident=seq_ident,
-        seq_covg=seq_covg,
+        bbh_results=bbh_results,
+        all_blast_hits=all_qry2ref_blast_hits,
     )
-    prokka_contig_cdss = [f for f in contig_features if f.type == 'CDS']
     with multiprocessing.Pool(processes=nproc) as pool:
         contig_features, ref_matched, coords_corrected = zip(*pool.map(
             mp_postprocess_feature,
@@ -155,8 +167,6 @@ def postprocess_contig(
     abinit_features_postprocessed_list, dropped_abinit_fragments = fissionfuser(
         contig_features,
         ref_annotation,
-        seq_ident=seq_ident,
-        seq_covg=seq_covg,
     )
 
     record.features = abinit_features_postprocessed_list
@@ -169,8 +179,8 @@ def postprocess_feature(
         feature,
         ref_annotation,
         ref_proteome,
-        seq_ident,
-        seq_covg,
+        bbh_results,
+        all_blast_hits,
 ):
     """
     general postprocessing of individual ab initio features.
@@ -179,8 +189,6 @@ def postprocess_feature(
     :param feature: AutarkicSeqFeature ab initio feature
     :param ref_annotation: dict of curated reference annotations
     :param ref_proteome: str file name of multi fasta reference amino acid sequences
-    :param seq_ident: int sequence identity percentage threshold for BLAST
-    :param seq_covg: int sequence alignment coverage percent threshold for BLAST
     """
     ref_matched = False
     coords_corrected = False
@@ -193,12 +201,8 @@ def postprocess_feature(
     if feature.type != 'CDS':
         return feature, ref_matched, coords_corrected
 
-    top_hit, low_covg, blast_hits = BLAST.reference_match(
-        SeqRecord(Seq(feature.qualifiers['translation'][0])),
-        subject=ref_proteome,
-        seq_ident=seq_ident,
-        seq_covg=seq_covg,
-    )
+    top_hit = bbh_results[feature.qualifiers['locus_tag'][0]]
+    blast_hits = all_blast_hits[feature.qualifiers['locus_tag'][0]]
 
     if top_hit:
         ref_matched = True
@@ -209,8 +213,6 @@ def postprocess_feature(
         feature_is_pseudo = pseudoscan(
             feature,
             ref_annotation[key_ref_gene(ref_id, ref_gene)],
-            seq_ident,
-            seq_covg,
             attempt_rescue=True,
             blast_hit_dict=blast_hits[ref_gene]
         )